project/backend/apps/Northwind/requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
ddtrace==0.57.3
dj-database-url==0.5.0
Django==3.2.11
django-cockroachdb==3.2.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
names==0.3.0
packaging==21.3
protobuf==3.19.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
python-json-logger==2.0.2
pytz==2021.3
requests==2.27.1
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.31
sqlalchemy-cockroachdb==1.4.3
sqlparse==0.4.2
tenacity==8.0.1
text-unidecode==1.3
typing-extensions==4.0.1
urllib3==1.26.8
zipp==3.7.0
